Create SelfReportCodeBox component
